Foundation: imperfect workaround for uint ambiguity problems

+++ implementations/MaxMSP/adc/jcom.adc.cpp
@@ -183,7 +183,7 @@ MaxErr AdcSetSampleRate(AdcPtr self, void* attr, AtomCount argc, AtomPtr argv)
 {
 	if (argc) {
 		TTUInt32 sr = atom_getlong(argv);
-		self->audioGraphObject->getUnitGenerator()->setAttributeValue(TT("sampleRate"), sr);
+		self->audioGraphObject->getUnitGenerator()->setAttributeValue(TT("sampleRate"), (uint)sr);
 	}
 	return MAX_ERR_NONE;
 }
@@ -206,7 +206,7 @@ MaxErr AdcSetVectorSize(AdcPtr self, void* attr, AtomCount argc, AtomPtr argv)
 {
 	if (argc) {
 		TTUInt32 vs = atom_getlong(argv);
-		self->audioGraphObject->getUnitGenerator()->setAttributeValue(TT("vectorSize"), vs);
+		self->audioGraphObject->getUnitGenerator()->setAttributeValue(TT("vectorSize"), (uint)vs);
 	}
 	return MAX_ERR_NONE;
 }
